use big decimal

+++ app/src/main/java/ru/bagrusss/revolutdemo/mappers/rates/RatesMapper.kt
@@ -1,8 +1,9 @@
 package ru.bagrusss.revolutdemo.mappers.rates
 
 import ru.bagrusss.revolutdemo.mappers.Mapper
+import java.math.BigDecimal
 
 /**
  * Created by bagrusss on 13.01.2020
  */
-interface RatesMapper : Mapper<Map<String, Double>, List<Pair<String, Double>>>
+interface RatesMapper : Mapper<Map<String, Double>, List<Pair<String, BigDecimal>>>
