+++ app/src/main/java/ru/bagrusss/revolutdemo/mappers/rates/RatesMapperImpl.kt
@@ -1,5 +1,6 @@
 package ru.bagrusss.revolutdemo.mappers.rates
 
+import java.math.BigDecimal
 import javax.inject.Inject
 
 /**
@@ -7,9 +8,9 @@ import javax.inject.Inject
  */
 class RatesMapperImpl @Inject constructor() : RatesMapper {
 
-    override fun map(from: Map<String, Double>): List<Pair<String, Double>> =
+    override fun map(from: Map<String, Double>): List<Pair<String, BigDecimal>> =
         from.map { (rate, cost) ->
-            rate to cost
+            rate to cost.toBigDecimal()
         }
 
 }
